Implement main menu
Massive refactor as a part of this commit, really cooked this one from
scratch instead of going for incremental improvements.

This is still not quite complete, but it includes:

New MenuMain state.
A UI framework with buttons and styling.
State switching.
Vertical rotation for mobile play with the new aspect ratio.
Innumerable changes to the code to make app states work out.